+++ INCLUDE/gssitype.h
@@ -4232,7 +4232,7 @@ void* __cdecl GSSicalloc(_In_ _CRT_GUARDOVERFLOW size_t _Count, _In_ _CRT_GUARDO
 #define PostMessageA GSSiPOSTMESSAGE 
 BOOL    WINAPI GSSiPOSTMESSAGE(HWND, UINT, WPARAM, LPARAM);
 #define	GlobalLock	GSSiGLOBALLOCK
-void LogMemAlloc (unsigned short MemID,long MemLen);
+void LogMemAlloc (int MemID,long MemLen);
 LPVOID GSSiGLOBALLOCK (HANDLE hglb);
 #define	GlobalSize	GSSiGLOBALSIZE
 DWORD GSSiGLOBALSIZE (HANDLE hglb);
